server/requirements-orchestrator.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
pydantic>=2.0
python-dotenv==1.0.1
# Mem0 + ChromaDB (L3 사실 추출 + 시맨틱 카드)
mem0ai>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
# BERTopic (주간 토픽 마이닝)
bertopic>=0.16.0
hdbscan>=0.8.33
umap-learn>=0.5.0
# STUMPY (일일 시계열 패턴)
stumpy>=1.12.0
numpy>=1.24.0
# Graphiti (월간 지식 그래프)
graphiti-core[kuzu]>=0.5.0
# NeuralProphet + TFT (일일 행동 예측)
neuralprophet>=0.8.0
pytorch-forecasting>=1.0.0
pandas>=2.0.0
